rename: host 도메인에 대한 api 연동 폴더로 묶기
